Moved all extensions to an individual file depending on which variable type they are interacting with. This prevents the code from looking like a rats nest even though there is a lot more files to look at. This makes more sense when some developer comes in and has to look at their code as they no longer need to goto Foundations and scroll like crazy (if they even know what Foundations means)
