+++ src/main/java/me/superpenguin/superglue/foundations/extensions/CommandSender.kt
@@ -1,0 +1,7 @@
+package me.superpenguin.superglue.foundations.extensions
+
+import org.bukkit.command.CommandSender
+
+/** Send a colored message to a player */
+fun CommandSender.send(msg: String, hex: Boolean = false) = sendMessage(msg.toColor(hex))
+
